change europa rent parameters to zero
substrate have removed the rent in pallet-contracts in newest substrate after this pr https://github.com/paritytech/substrate/pull/9669. 

but we do not wanna update current substrate to new version, for we are waiting substrate to release 4.0.0.

thus, we just change the rent parameters to zero, it will cause the same result compared with removed rent pallet-contracts.